requirements.txt
paho-mqtt
keyboard
psutil
